requirements.txt
absl-py==1.1.0
ale-py==0.7.5
appnope==0.1.3
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
baselines @ file:///Users/sachingiri/sachin/projects/gym-car/baselines
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.2.2
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
grpcio==1.46.3
gym==0.15.7
gym-notices==0.0.7
h5py==3.7.0
idna==3.3
importlib-resources==5.8.0
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-rl==0.4.2
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pygame==2.1.2
pyglet==1.5.0
Pygments==2.12.0
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six==1.16.0
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
termcolor==1.1.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
